backend/requirements.txt
APScheduler==3.8.0
boto3==1.17.17
botocore==1.20.112
cacheout==0.13.1
CacheControl==0.12.6
celery==5.0.5
channels==3.0.3
Django==3.2.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-eventstream==4.2.0
django-extensions==3.1.0
django-filter==2.4.0
django-fsm==2.7.1
django-model-utils==4.1.1
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-solo==1.1.5
django-storages==1.11.1
django-templated-mail==1.1.1
django-tenants==3.3.2
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
drf-yasg==1.20.0
firebase-admin==5.0.1
google-api-core==1.31.3
google-api-python-client==2.22.0
ipython==7.20.0
phonenumberslite==8.12.33
Pillow==8.1.0
pyfcm==1.5.4
psycopg2==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
r-django-jet
django-recaptcha
django-axes
django-password-validators
# django-user-accounts
django-select2
